Support Rust 1.0.0 and document it
I was too lazy to figure out which is the real MSRV so I just tried
1.0.0 and it turned out it only needed the previous change, which was
useful anyway, and then just a trivial one-line change. While the change
arguably worsens the style by a tiny little bit it's well worth the meme
value of supporting 1.0.0. A `clippy.toml` was added to not annoy
developers about it.

The MSRV was also undocumented, so this adds a documentation as well.
The `rust-version` field is not added to not break the versions of rust
that were banning it as "unstable".